// ui/display/bank-logo/types.ts
export type FipId =
  | "ICICI-FIP"
  | "Hutatma_FIPLIVE"
  | "AUBank-FIP"
  | "IDFCFirstBank-FIP"
  | "KotakMahindraBank-FIP"
  | "idbibank-fip"
  | "fip@nsdl"
  | "UBI-FIP"
  | "UCOB-FIP"
  | "kfinmf-fip"
  | "CUBFIP"
  | "MVAFL_live"
  | "KBL-FIP"
  | "IOB-FIP"
  | "OGB-FIP"
  | "fiplive@canarabank"
  | "fiplive@fincarebank"
  | "IBFIP"
  | "CAMSRTAFIP"
  | "BARBFIP"
  | "PSFIP"
  | "YESB-FIP"
  | "HDFC-FIP"
  | "CENTRALFIP"
  | "sbi-fip"
  | "AXIS001"
  | "BOM_FIP"
  | "FDRLFIPPROD"
  | "CGGBFIP0001"
  | "CDSLFIP"
  | "PNB-FIP"
  | "KarurVysyaBank-FIP"
  | "BOI-FIP"
  | "MGBFIP"
  | "fiplive@indusind";